+++ src/run.ts
@@ -10,6 +10,7 @@ import {
 } from "./shell-utils";
 import { getAppPort, getAutoreloadPort } from "./port-settings";
 import * as winreg from "winreg";
+import { getPositronPreferredRuntime } from "./extension-api-utils/extensionHost";
 
 const DEBUG_NAME = "Debug Shiny app";
 
@@ -344,7 +345,56 @@ function getExtensionPath(): string | undefined {
 }
 
 async function getRBinPath(bin: string): Promise<string> {
-  return getRPathFromEnv(bin) || (await getRPathFromWindowsReg(bin)) || "";
+  return (
+    (await getRPathFromPositron(bin)) ||
+    getRPathFromConfig(bin) ||
+    getRPathFromEnv(bin) ||
+    (await getRPathFromWindowsReg(bin)) ||
+    ""
+  );
+}
+
+async function getRPathFromPositron(bin: string): Promise<string> {
+  const runtimeMetadata = await getPositronPreferredRuntime("r");
+  if (!runtimeMetadata) {
+    return "";
+  }
+
+  console.log(`[shiny] runtimeMetadata: ${JSON.stringify(runtimeMetadata)}`)
+
+  const runtimePath = runtimeMetadata.runtimePath;
+  if (!runtimePath) {
+    return "";
+  }
+
+  const { platform } = process;
+  const fileExt = platform === "win32" ? ".exe" : "";
+  return path_join(path_dirname(runtimePath), bin + fileExt);
+}
+
+function getRPathFromConfig(bin: string): string {
+  const { platform } = process;
+  const fileExt = platform === "win32" ? ".exe" : "";
+  let osType: string;
+
+  switch (platform) {
+    case "win32":
+      osType = "windows";
+      break;
+    case "darwin":
+      osType = "mac";
+      break;
+    default:
+      osType = "linux";
+  }
+
+  const rPath = vscode.workspace
+    .getConfiguration("r.rpath")
+    .get(osType, undefined);
+
+  console.log(`[shiny] rPath: ${rPath}`);
+
+  return rPath ? path_join(path_dirname(rPath), bin + fileExt) : "";
 }
 
 function getRPathFromEnv(bin: string = "R"): string {